pyproject.toml
[tool.poetry]
name = "analyze-eth-events"
version = "0.1.0"
description = ""
authors = ["Alexey Useless, Sergey Boogerwooger"]

[tool.poetry.dependencies]
python = "^3.9"
web3 = "^5.26.0"
psycopg2-binary = "^2.9.3"
python-dotenv = "^0.19.2"
argparse = "^1.4.0"
bitstring = "^3.1.9"
eth-account = "^0.5.6"
numpy = "^1.22.1"
jsonrpcserver = "^5.0.6"
fastapi = "^0.73.0"
uvicorn = "^0.17.5"
requests = "^2.27.1"
ethereum = "^2.3.2"
pyTelegramBotAPI = "^4.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
